ai_blog/pyproject.toml
[project]
name = "blog-proj-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.112.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.6",
    "motor>=3.5.1",
    "langchain>=0.2.14",
    "langchain-openai>=0.1.22",
    "jinja2>=3.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
